Clean up tests and add a few more
